refactor(agent): 从 chat.py 拆出 agent_turn.py 分离 WS 通信与 Agent 编排
api/routes/chat.py 原混合两个职责：WebSocket 传输层和 Agent 对话编排。
拆分为两个文件，职责清晰：

api/agent_turn.py（新增，493 行）：Agent 对话编排
  - run_agent_turn()（原 _run_agent_turn，改为公开）
  - _stream_turn() / _get_final_answer()
  - _inject_cancel_tool_messages()
  - _get_provider_context() / _get_session_messages()
  - _calculate_context_usage()

api/routes/chat.py（精简为 127 行）：WebSocket 传输层
  - websocket_chat() — accept/receive/disconnect、消息派发
  - _resume_sub_agent() — 重连恢复

server.py 无改动，chat 模块仍在、router 仍在。
